% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{cbind_unique}
\alias{cbind_unique}
\title{Bind Data Frames and Remove Duplicates}
\usage{
cbind_unique(x, y, cols)
}
\arguments{
\item{x}{A data frame from which duplicates will be removed.}

\item{y}{A data frame containing the columns to bind into \code{x}.}

\item{cols}{A character vector of column names to use for binding.}
}
\value{
A data frame with the bound columns and duplicates removed.
}
\description{
This function binds two data frames and removes duplicates from the first data frame based on the columns in the second data frame.
}
\examples{
\dontrun{
  cbind_unique(df1, df2, c("col1", "col2"))
}
}
