% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora_to_ggwalter_lieth.R
\name{check_graph}
\alias{check_graph}
\title{Check Graph}
\usage{
check_graph(graph_object)
}
\arguments{
\item{graph_object}{The graph object to be checked.}
}
\value{
The recorded plot object or NULL if an error occurs.
}
\description{
Records the plot if \code{graph_object} is NULL and returns the graph object of class "recordedplot". Applicable to base graphs only.
}
\examples{
\dontrun{
  check_graph(my_graph)
}
}
