% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_character_matrix.R
\name{convert_to_character_matrix}
\alias{convert_to_character_matrix}
\title{Convert Data to Character Matrix}
\usage{
convert_to_character_matrix(
  data,
  format_decimal_places = TRUE,
  decimal_places,
  is_scientific = FALSE,
  return_data_frame = TRUE,
  na_display = NULL,
  check.names = TRUE
)
}
\arguments{
\item{data}{The input data to be converted.}

\item{format_decimal_places}{A logical value indicating whether decimal places should be formatted. Default is \code{TRUE}.}

\item{decimal_places}{A numeric vector specifying the number of decimal places for each column. If not provided, it uses the function \code{get_default_significant_figures} to determine the number of decimal places.}

\item{is_scientific}{A logical vector indicating whether scientific notation should be used for each column. Default is \code{FALSE}.}

\item{return_data_frame}{A logical value indicating whether the result should be returned as a data frame. Default is \code{TRUE}.}

\item{na_display}{A character value specifying how missing values should be displayed. If not provided, missing values are left as is.}

\item{check.names}{A logical value indicating whether column names should be checked and modified if necessary. Default is `TRUE'.}
}
\value{
The converted data as a character matrix or data frame, depending on the value of \code{return_data_frame}.
}
\description{
This function converts the input data to a character matrix, with options for formatting decimal places and handling missing values.
}
\examples{
data <- data.frame(A = c(1.234, 5.678), B = c(10.123, 20.456))
converted <- convert_to_character_matrix(data)
print(converted)

# example with SFC data
points <- sf::st_sfc(sf::st_point(c(1, 2)), sf::st_point(c(3, 4)))
sf_data <- data.frame(id = c(1, 2))
sf_data$geometry <- points
class(sf_data$geometry) <- c("sfc", class(sf_data$geometry))  # Ensure "sfc" class
# Run function
result <- convert_to_character_matrix(sf_data)

}
\seealso{
\code{\link{get_default_significant_figures}}
}
