% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{convert_to_list}
\alias{convert_to_list}
\title{Convert Character to List of Numeric Vectors}
\usage{
convert_to_list(x)
}
\arguments{
\item{x}{A character string to be converted.}
}
\value{
A list of numeric vectors.
}
\description{
Converts a character string to a list of numeric vectors.
}
\examples{
\dontrun{
  convert_to_list("c(1,2,3)")
}
}
