% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rankings_list.R
\name{create_rankings_list}
\alias{create_rankings_list}
\title{Create a List of Rankings for Multiple Traits}
\usage{
create_rankings_list(
  data,
  traits = traits,
  id = "id",
  variety = "variety",
  group = FALSE
)
}
\arguments{
\item{data}{A data frame containing ranking information in wide format.}

\item{traits}{A character vector of column names corresponding to traits to be ranked.}

\item{id}{A string specifying the column name for individual identifiers.}

\item{variety}{A string specifying the column name for item (e.g., variety) labels.}

\item{group}{Logical. Whether to group rankings (e.g., for grouped ranking designs). Default is \code{FALSE}.}
}
\value{
A named list of ranking objects (one for each trait) generated using \code{gosset::rank_tricot2}.
}
\description{
Converts a wide-format dataset with ranking values into a list of ranking objects,
one per trait, after filtering out any rows with missing values.
}
\examples{
library(gosset)
data(cassava)
cassava_by_id_variety <- pivot_tricot(cassava)
traits <- c("colour", "firmness", "odour", "overall", "stretchability", "taste", "mouldability", "smoothness")
rankings_list <- create_rankings_list(data = cassava_by_id_variety, traits, group = FALSE)
grouped_rankings_list <- create_rankings_list(data = cassava_by_id_variety, traits, group = TRUE)
}
