% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora to ggwalter_lieth.R
\name{cumulative_inventory}
\alias{cumulative_inventory}
\title{Cumulative Inventory}
\usage{
cumulative_inventory(data, station = NULL, from, to)
}
\arguments{
\item{data}{The input data frame.}

\item{station}{An optional column name indicating the station. Defaults to NULL.}

\item{from}{The start period column name.}

\item{to}{The end period column name.}
}
\value{
A data frame with added cumulative counts.
}
\description{
Adds a cumulative count of missing values by period (and station if specified) to the data.
}
\examples{
\dontrun{
  cumulative_inventory(my_data, "station", "start_date", "end_date")
}
}
