% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{decimals}
\alias{decimals}
\title{Decimal Representation}
\usage{
decimals(x)
}
\arguments{
\item{x}{A character vector containing fractions expressed as strings.}
}
\value{
A numeric vector containing the decimal equivalents of the
supplied fractions.
}
\description{
Converts fractional character strings into decimal values.
}
\details{
Fraction strings are evaluated and converted to their corresponding
decimal values.

This function is intended as the inverse operation of
\code{\link{fractions}}.
}
\examples{
decimals(c("3/4", "23/10", "7/25"))
# Returns c(0.75, 2.3, 0.28)

fractions(decimals(c("3/4", "23/10")))
# Returns c("3/4", "23/10")

}
\seealso{
\code{\link{fractions}}
}
