% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{digitsqu}
\alias{digitsqu}
\title{Squares of Digits}
\usage{
digitsqu(x)
}
\arguments{
\item{x}{A numeric or integer vector.}
}
\value{
A list containing the squared digits for each input value.
}
\description{
Computes the square of each digit in an integer.
}
\details{
Each element is split into its constituent digits and each digit is
squared individually.
}
\examples{
digitsqu(c(8, 23, 471))
# Returns:
# 64
# c(4, 9)
# c(16, 49, 1)

}
