% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{digitsum}
\alias{digitsum}
\title{Sum of Digits}
\usage{
digitsum(x)
}
\arguments{
\item{x}{A numeric or integer vector.}
}
\value{
A numeric vector containing the sum of digits for each element.
}
\description{
Computes the sum of the digits for each integer value.
}
\details{
Each number is converted to its individual digits and these digits
are summed.

This function is intended to provide functionality similar to
\code{DescTools::DigitSum()}, although the behaviour may differ for
}
