% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gt_table.R
\name{format_gt_table}
\alias{format_gt_table}
\title{Format gt Tables}
\usage{
format_gt_table(df, ...)
}
\arguments{
\item{df}{A data frame containing the summary data.}

\item{...}{Additional options to read into gt()}
}
\value{
A \code{gt} table with formatted styling and an automatically generated title.
}
\description{
Add Description
}
