% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{frac10}
\alias{frac10}
\title{Convert Decimal to Fraction}
\usage{
frac10(x)
}
\arguments{
\item{x}{A numeric value representing the decimal to convert.}

\item{den}{An integer specifying the denominator for the fraction. (Only for \code{frac_den}.)}
}
\value{
A character string representing the fraction.
}
\description{
Converts a decimal value into a fraction with a specified denominator or common denominators (10, 20, or 100).
}
\examples{
# Convert decimals to fractions
frac10(0.75)  # "8/10"
frac20(0.25)  # "5/20"
frac100(0.123) # "12/100"
frac_den(0.333, 3) # "1/3"

}
