% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{fractions}
\alias{fractions}
\title{Fraction Representation}
\usage{
fractions(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector containing fractional representations of
the supplied values.
}
\description{
Converts decimal values to fractional representations.
}
\details{
This function is a wrapper around \code{MASS::fractions()} and returns
the results as character strings rather than fraction objects.
}
\examples{
fractions(c(0.75, 2.3, 0.28))
# Returns c("3/4", "23/10", "7/25")

}
\seealso{
\code{\link{decimals}}
}
