% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_tables.R
\name{generate_summary_tables}
\alias{generate_summary_tables}
\title{Generate Summary Tables}
\usage{
generate_summary_tables(table_data, rm_summary_title = FALSE, ...)
}
\arguments{
\item{table_data}{A data frame containing the summary data, including a \code{summary-variable} column.}

\item{rm_summary_title}{Optional boolean to remove the summary label from the title. Only if there is one summary given, and it is a \code{summary-variable} type. Default \code{FALSE}.}

\item{...}{Additional options to read into gt()}
}
\value{
A \code{gt} table with formatted styling and an automatically generated title.
}
\description{
This function creates a styled summary table using \code{gt}, formatted with
a title extracted from the \code{summary-variable} column. The table is designed
to display summarized data while dropping the \code{summary-variable} column.
}
