% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora to ggwalter_lieth.R
\name{get_vignette}
\alias{get_vignette}
\title{Get Vignette}
\usage{
get_vignette(package = NULL, lib.loc = NULL, all = TRUE)
}
\arguments{
\item{package}{The name of the package. Defaults to NULL.}

\item{lib.loc}{The library location. Defaults to NULL.}

\item{all}{Logical indicating whether to list all available vignettes. Defaults to TRUE.}
}
\value{
A character vector with the vignette information.
}
\description{
Retrieves the vignette information for a specified package.
}
\examples{
\dontrun{
  get_vignette("dplyr")
}
}
