% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora to ggwalter_lieth.R
\name{getExample}
\alias{getExample}
\title{Get Example}
\usage{
getExample(
  topic,
  package = NULL,
  lib.loc = NULL,
  character.only = TRUE,
  give.lines = FALSE,
  local = FALSE,
  echo = TRUE,
  verbose = getOption("verbose"),
  setRNG = FALSE,
  ask = getOption("example.ask"),
  prompt.prefix = abbreviate(topic, 6),
  run.dontrun = FALSE,
  run.donttest = interactive()
)
}
\arguments{
\item{topic}{The topic for which to retrieve the example.}

\item{package}{The name of the package containing the topic. Defaults to NULL.}

\item{lib.loc}{The library location. Defaults to NULL.}

\item{character.only}{Logical indicating whether the topic is specified as a character string. Defaults to TRUE.}

\item{give.lines}{Logical indicating whether to return the example code as lines of text. Defaults to FALSE.}

\item{local}{Logical indicating whether to evaluate the example locally. Defaults to FALSE.}

\item{echo}{Logical indicating whether to echo the example code. Defaults to TRUE.}

\item{verbose}{Logical indicating whether to print verbose output. Defaults to the value of the "verbose" option.}

\item{setRNG}{Logical indicating whether to set the random number generator. Defaults to FALSE.}

\item{ask}{Logical indicating whether to ask before evaluating the example. Defaults to the value of the "example.ask" option.}

\item{prompt.prefix}{A prefix for the prompt. Defaults to an abbreviated version of the topic.}

\item{run.dontrun}{Logical indicating whether to run examples marked with \dontrun. Defaults to FALSE.}

\item{run.donttest}{Logical indicating whether to run examples marked with \donttest. Defaults to the value of \code{interactive()}.}
}
\value{
The example code as a character string if \code{give.lines} is TRUE, otherwise prints the example code.
}
\description{
Retrieves and displays the example code for a specified topic from a package.
}
\examples{
\dontrun{
  getExample("filter", "dplyr")
}
}
