% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPass.R
\name{getPass}
\alias{getPass}
\title{Password Input (from \code{getPass} v0.2.2)}
\usage{
getPass(msg = "PASSWORD: ", noblank = FALSE, forcemask = FALSE)
}
\arguments{
\item{msg}{The message to enter into the R session before prompting
for the masked input. This can be any single string, possibly including
a "blank" (\code{""}); see the \code{noblank} argument.}

\item{noblank}{Logical; should blank passwords (\code{""}) be banned? By default,
they are allowed, except with RStudio where they are always banned.}

\item{forcemask}{Logical; should the function stop with an error if masking
is not supported? If \code{FALSE}, the function will default
to use \code{readline()} with a warning message that the
input is not masked, and otherwise will stop with an error.
See the Details section for more information.}
}
\value{
If input is provided, then that is returned. If the user cancels
(e.g., cancel button on RStudio or ctrl+c in the terminal), then
\code{NULL} is returned.
}
\description{
A password reader function adapted from version 0.2.2 of the
\strong{getPass} package. Like R's \code{readline()}, but the
user-typed input text is not printed to the screen.
}
\details{
This function is based on the implementation from getPass 0.2.2,
originally authored by Drew Schmidt. It provides a simple way to
securely capture user input without displaying it on the screen.

Masking (i.e., not displaying the literal typed text as input) is supported
on most, but not all platforms. It is supported in RStudio, provided you
have a suitable version of the GUI. It should also work in the terminal
on any major OS. Additionally, it will work in any environment where
the tcltk package is available (e.g., Windows with RGui).
Notably, this will not work with Emacs; passwords can be read, but
there will be no masking.

In the terminal, the maximum length for input is 255 characters.
Additionally, messages printed to the terminal (including the "*" masking)
are printed to stderr.
}
\note{
This function is directly adapted from version 0.2.2 of the \code{getPass} package.
Full credit to Drew Schmidt for the original implementation.
}
\examples{
\dontrun{
# Basic usage
getPass::getPass()

# Get password with a custom message
getPass::getPass("Enter the password: ")
}

}
\references{
Schmidt, D. (2017). \emph{Guide to the getPass Package}.
R package version 0.2.2. Retrieved from \url{https://CRAN.R-project.org/package=getPass}
}
