% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora_to_ggwalter_lieth.R
\name{getRowHeadersWithText}
\alias{getRowHeadersWithText}
\title{Get Row Headers with Text}
\usage{
getRowHeadersWithText(data, column, searchText, ignore_case, use_regex)
}
\arguments{
\item{data}{The input data frame.}

\item{column}{The name of the column to search.}

\item{searchText}{The text to search for.}

\item{ignore_case}{Logical indicating whether to ignore case. Defaults to TRUE.}

\item{use_regex}{Logical indicating whether to use regular expressions. Defaults to FALSE.}
}
\value{
A character vector of row headers where the search text is found.
}
\description{
Retrieves the row headers of a data frame where the specified column contains the search text.
}
\examples{
\dontrun{
  getRowHeadersWithText(my_data, "column1", "search text", TRUE, FALSE)
}
}
