% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora_to_ggwalter_lieth.R
\name{ggwalter_lieth}
\alias{ggwalter_lieth}
\title{Generate Walter-Lieth Plot}
\usage{
ggwalter_lieth(
  data,
  month,
  station = NULL,
  p_mes,
  tm_max,
  tm_min,
  ta_min,
  station_name = "",
  alt = NA,
  per = NA,
  pcol = "#002F70",
  tcol = "#ff0000",
  pfcol = "#9BAEE2",
  sfcol = "#3C6FC4",
  shem = FALSE,
  p3line = FALSE,
  ...
)
}
\arguments{
\item{data}{The input data frame.}

\item{month}{The column name for months.}

\item{station}{An optional column name for station. Defaults to NULL.}

\item{p_mes}{The column name for precipitation measurements.}

\item{tm_max}{The column name for maximum temperature.}

\item{tm_min}{The column name for minimum temperature.}

\item{ta_min}{The column name for average temperature.}

\item{station_name}{The name of the station. Defaults to an empty string.}

\item{alt}{The altitude. Defaults to NA.}

\item{per}{The period. Defaults to NA.}

\item{pcol}{The color for precipitation lines. Defaults to "#002F70".}

\item{tcol}{The color for temperature lines. Defaults to "#ff0000".}

\item{pfcol}{The fill color for probable freeze areas. Defaults to "#9BAEE2".}

\item{sfcol}{The fill color for sure freeze areas. Defaults to "#3C6FC4".}

\item{shem}{Logical indicating whether the station is in the southern hemisphere. Defaults to FALSE.}

\item{p3line}{Logical indicating whether to plot the precipitation/3 line. Defaults to FALSE.}

\item{...}{Additional arguments to be passed to ggplot2 functions.}
}
\value{
A ggplot2 object with the Walter-Lieth climate diagram.
}
\description{
Generates Walter-Lieth climate diagrams using ggplot2.
}
\examples{
\dontrun{
  ggwalter_lieth(my_data, "Month", NULL, "Precipitation", "MaxTemp", "MinTemp", "AvgTemp")
}
}
