% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in_operator.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not-in Operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{A value or vector to check.}

\item{table}{A vector to check against.}
}
\value{
A logical vector indicating if \code{x} is not in \code{table}.
}
\description{
This operator returns \code{TRUE} if the left-hand side value is not in the right-hand side vector.
}
\examples{
5 \%notin\% c(1, 2, 3, 4) # TRUE
"a" \%notin\% c("a", "b", "c") # FALSE

}
