% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscontainPartialValueLabel.R
\name{is.containPartialValueLabel}
\alias{is.containPartialValueLabel}
\title{Check for Partial Value Labels}
\usage{
is.containPartialValueLabel(x)
}
\arguments{
\item{x}{A variable to check for partial value labels.}
}
\value{
A logical value. Returns \code{TRUE} if the variable contains partial value labels, otherwise \code{FALSE}.
}
\description{
Checks if the variable contains partial value labels (some values labeled, others not).
}
\examples{
# Example with partially labeled variable
#is.containPartialValueLabel(x)

}
