% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_names_function.R
\name{make_names_pattern}
\alias{make_names_pattern}
\title{Create a \code{names_pattern} for \code{tidyr::pivot_longer()}}
\usage{
make_names_pattern(items, num)
}
\arguments{
\item{items}{A character vector of column names to be grouped.}

\item{num}{The number of groups into which \code{items} should be divided.}
}
\value{
A character string containing a regular expression for use with
\code{tidyr::pivot_longer(names_pattern = ...)}.
}
\description{
Creates a regular expression suitable for the \code{names_pattern} argument of
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}} from a vector of column names and the number of
groups represented by those columns.
}
\details{
The columns in \code{items} are assumed to be ordered so that corresponding
variables occur in the same position within each group. The columns are
divided into \code{num} groups using \code{\link[=split_items_in_groups]{split_items_in_groups()}}. The common prefix
of each group is used as the \code{.value} component, while the part of the
names following the prefix is used as the \code{variable} component.
}
\examples{
make_names_pattern(
  c("X1", "X2", "X3", "X4",
    "Y1", "Y2", "Y3", "Y4"),
  num = 2
)
# "(X|Y)(1|2|3|4)"

make_names_pattern(
  c("height_male", "height_female",
    "weight_male", "weight_female"),
  num = 2
)
# "(height_|weight_)(male|female)"
}
