% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_consecutive_sum.R
\name{max_consecutive_sum}
\alias{max_consecutive_sum}
\title{Compute the Maximum Consecutive Sum}
\usage{
max_consecutive_sum(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the maximum consecutive sum.
}
\description{
This function calculates the maximum sum of consecutive elements in a numeric vector.
}
\details{
This function relies on \code{consecutive_sum()}, assuming it computes
consecutive sums with an initial value of 0. The maximum of these sums is returned.
}
\examples{
x <- c(1, -2, 3, 4, -1, 2, 1, -5, 4)
max_consecutive_sum(x)

}
\seealso{
\code{\link{consecutive_sum}}
}
