% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{monitor_memory}
\alias{monitor_memory}
\title{Monitor Memory Usage}
\usage{
monitor_memory()
}
\value{
A numeric value representing the memory usage in MB.
}
\description{
Monitors and returns the current memory usage in megabytes (MB).
}
\examples{
# Check memory usage
monitor_memory()

}
