% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{pascal}
\alias{pascal}
\title{Pascal Triangle Coefficients}
\usage{
pascal(x)
}
\arguments{
\item{x}{A vector of non-negative integers.}
}
\value{
A list containing the binomial coefficients for each supplied
integer.
}
\description{
Generates binomial coefficients corresponding to rows of Pascal's Triangle.
}
\details{
For each value \code{n}, the function calculates:

\deqn{\binom{n}{0}, \binom{n}{1}, \ldots, \binom{n}{n}}

These values correspond to the rows of Pascal's Triangle.
}
\examples{
pascal(c(1, 2, 3, 4))

# Returns:
# c(1, 1)
# c(1, 2, 1)
# c(1, 3, 3, 1)
# c(1, 4, 6, 4, 1)

}
