% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{plot_network}
\alias{plot_network}
\title{Visualise a ranking network using \code{ggplot2} and \code{ggraph}}
\usage{
plot_network(object, fluctuate_widths = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{rankings} or \code{grouped_rankings}.}

\item{fluctuate_widths}{Logical; if \code{TRUE}, edge widths reflect edge weights (e.g., strength or frequency of preference).}

\item{...}{Additional arguments passed to \code{ggraph::ggraph} functions.}
}
\value{
A \code{ggplot2} network plot representing the ranking structure.
}
\description{
Visualise a ranking network using \code{ggplot2} and \code{ggraph}
}
\examples{
library("PlackettLuce")

R = matrix(c(1, 2, 0, 0,
             4, 1, 2, 3,
             2, 4, 1, 3,
             1, 2, 3, 0,
             2, 1, 3, 0,
             1, 0, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) = c("apple", "banana", "orange", "pear")
R = as.rankings(R)

plot_network(R, fluctuate_widths = TRUE)
}
