% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora_to_ggwalter_lieth.R
\name{prepare_walter_lieth}
\alias{prepare_walter_lieth}
\title{Prepare Data for Walter-Lieth Plot}
\usage{
prepare_walter_lieth(data, month, tm_min, ta_min)
}
\arguments{
\item{data}{The input data frame.}

\item{month}{The column name for months.}

\item{tm_min}{The column name for minimum temperature.}

\item{ta_min}{The column name for average temperature.}
}
\value{
A list of data frames and plots prepared for Walter-Lieth climate diagrams.
}
\description{
Prepares data for Walter-Lieth climate diagrams.
}
\examples{
\dontrun{
  prepare_walter_lieth(my_data, "Month", "MinTemp", "AvgTemp")
}
}
