% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{process_html_object}
\alias{process_html_object}
\title{Process Individual HTML Object}
\usage{
process_html_object(html_object)
}
\arguments{
\item{html_object}{The HTML object to be processed.}
}
\value{
The HTML object or the file name if saved as an HTML file.
}
\description{
Displays the HTML object in the R viewer. If the viewer is not available, saves the object as a file in the temporary folder.
}
\examples{
\dontrun{
  process_html_object(my_html_object)
}
}
