% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_variable_levels.R
\name{return_variable_levels}
\alias{return_variable_levels}
\title{Generate JavaScript Sorting Function for Categorical Variables}
\usage{
return_variable_levels(data)
}
\arguments{
\item{data}{A data frame containing categorical variables (factors).}
}
\value{
A character string containing a JavaScript function that defines
custom sorting orders for categorical variables.
}
\description{
This function identifies categorical variables (factors) in a given dataset
and generates a JavaScript sorting function that specifies the sorting order
for each categorical variable based on its levels.
}
\details{
This function is for use in the \verb{Pivot Tables} dialog
}
\examples{
df <- data.frame(
  category1 = factor(c("low", "medium", "high"), levels = c("low", "medium", "high")),
  category2 = factor(c("A", "B", "C"), levels = c("A", "B", "C"))
)
js_sort_function <- return_variable_levels(df)
cat(js_sort_function)
}
