% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{set_library_paths}
\alias{set_library_paths}
\title{Set Library Paths}
\usage{
set_library_paths(version)
}
\arguments{
\item{version}{A character string specifying the version}
}
\value{
This function does not return a value but modifies the \code{.libPaths()} environment variable.
}
\description{
This function updates the R library paths by adding a new library path at the beginning of the existing paths.
If there are more than two library paths, it retains only the first and third paths (if they exist).
}
