% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{time_operation}
\alias{time_operation}
\title{Time an Operation}
\usage{
time_operation(expr)
}
\arguments{
\item{expr}{An R expression to evaluate and time.}
}
\value{
Prints the time taken for the operation.
}
\description{
Measures and prints the time taken to execute an expression.
}
\examples{
# Time a simple operation
time_operation({ Sys.sleep(1); mean(1:100) })

}
