% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{view_graph_object}
\alias{view_graph_object}
\title{View Graph Object}
\usage{
view_graph_object(graph_object)
}
\arguments{
\item{graph_object}{The graph object to be displayed.}
}
\value{
The graph object or the file name if saved as an image.
}
\description{
Displays the graph object in the R viewer. If the viewer is not available, saves the object as a file in the temporary folder.
}
\examples{
\dontrun{
  view_graph_object(my_graph)
}
}
