% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora to ggwalter_lieth.R
\name{view_html_object}
\alias{view_html_object}
\title{View HTML Object}
\usage{
view_html_object(html_object)
}
\arguments{
\item{html_object}{The HTML object to be displayed.}
}
\value{
The HTML object or the file name if saved as an HTML file.
}
\description{
Displays the HTML object in the R viewer. If the viewer is not available, saves the object as a file in the temporary folder.
}
\examples{
\dontrun{
  view_html_object(my_html)
}
}
