% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{view_object}
\alias{view_object}
\title{View Object}
\usage{
view_object(data_book_object)
}
\arguments{
\item{data_book_object}{A data book object containing the object to be displayed and its format.}
}
\value{
The file name if the object is saved to a file, otherwise NULL.
}
\description{
Displays the object stored in a data book object.
}
\examples{
\dontrun{
  view_object(my_data_book)
}
}
