% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{view_object_data}
\alias{view_object_data}
\title{View Object Data}
\usage{
view_object_data(object, object_format = NULL)
}
\arguments{
\item{object}{The object to be displayed.}

\item{object_format}{The format in which to display the object ("image", "text", "html"). Defaults to NULL.}
}
\value{
A file name if the object is saved to a file, otherwise NULL.
}
\description{
Displays the given object in a specified format. If no format is provided, the object is printed.
}
\examples{
\dontrun{
  view_object_data(my_object, "text")
}
}
