% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_functions.R
\name{view_text_object}
\alias{view_text_object}
\title{View Text Object}
\usage{
view_text_object(text_object)
}
\arguments{
\item{text_object}{The text object to be displayed.}
}
\value{
The text object or the file name if saved as a text file.
}
\description{
Displays the text object in the R viewer. If the viewer is not available, saves the object as a file in the temporary folder.
}
\examples{
\dontrun{
  view_text_object(my_text)
}
}
