% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora_to_ggwalter_lieth.R
\name{WB_evaporation}
\alias{WB_evaporation}
\title{Calculate Evaporation for Water Balance}
\usage{
WB_evaporation(water_balance, frac, capacity, evaporation_value, rain)
}
\arguments{
\item{water_balance}{The current water balance.}

\item{frac}{The fraction of capacity.}

\item{capacity}{The total capacity.}

\item{evaporation_value}{The value of evaporation.}

\item{rain}{The amount of rain.}
}
\value{
The calculated evaporation.
}
\description{
Calculates the evaporation based on water balance, fraction of capacity, and other parameters.
}
\examples{
\dontrun{
  WB_evaporation(100, 0.5, 200, 10, 5)
}
}
