% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpora to ggwalter_lieth.R
\name{write_weather_data}
\alias{write_weather_data}
\title{Write Weather Data to File}
\usage{
write_weather_data(
  year,
  month,
  day,
  rain,
  mn_tmp,
  mx_tmp,
  missing_code,
  output_file
)
}
\arguments{
\item{year}{A vector of years.}

\item{month}{A vector of months.}

\item{day}{A vector of days.}

\item{rain}{A vector of rainfall values.}

\item{mn_tmp}{A vector of minimum temperatures.}

\item{mx_tmp}{A vector of maximum temperatures.}

\item{missing_code}{The code to use for missing values.}

\item{output_file}{The name of the output file.}
}
\value{
None. Writes the data to a file.
}
\description{
Writes weather data to a text file with specified missing value codes.
}
\examples{
\dontrun{
  write_weather_data(2020, 1:12, 1:31, rain_data, min_temp, max_temp, -99, "weather.txt")
}
}
