% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tricot.R
\name{pivot_tricot}
\alias{pivot_tricot}
\title{Pivot Tricot Data}
\usage{
pivot_tricot(
  data = NULL,
  data_id_col = "id",
  data_id_variety_trait = NULL,
  data_id_variety_trait_id_col = "id",
  variety_col = NULL,
  trait_col = NULL,
  rank_col = NULL,
  option_cols = c("option_a", "option_b", "option_c"),
  possible_ranks = c("A", "B", "C"),
  trait_good = "_pos",
  trait_bad = "_neg",
  na_value = "Not observed"
)
}
\arguments{
\item{data}{A data frame containing option columns and trait ranking indicators (e.g. columns ending in \verb{_pos} or \verb{_neg}).}

\item{data_id_col}{The name of the ID column in \code{data}. Default is \code{"id"}.}

\item{data_id_variety_trait}{A data frame containing ID, variety, trait, and rank columns.}

\item{data_id_variety_trait_id_col}{The name of the ID column in \code{data_id_variety_trait}. Default is \code{"id"}.}

\item{variety_col}{The name of the variety column in \code{data_id_variety_trait}.}

\item{trait_col}{The name of the trait column in \code{data_id_variety_trait}.}

\item{rank_col}{The name of the rank column in \code{data_id_variety_trait}.}

\item{option_cols}{A character vector of column names in \code{data} that store the option labels (e.g. \code{"option_a"}, \code{"option_b"}, \code{"option_c"}).}

\item{possible_ranks}{A character vector mapping to \code{option_cols}, indicating rank labels (e.g. \code{"A"}, \code{"B"}, \code{"C"}).}

\item{trait_good}{A suffix string (e.g. \code{"_pos"}) used to identify columns in \code{data} where a variety is "preferred" for a trait. Default is \code{"_pos"}.}

\item{trait_bad}{A suffix string (e.g. \code{"_neg"}) used to identify columns in \code{data} where a variety is "not preferred" for a trait. Default is \code{"_neg"}.}

\item{na_value}{A value used to indicate when a trait was not observed or is missing. Default is \code{"Not observed"}.}
}
\value{
A tibble with the transformed Tricot data, containing columns:
\itemize{
\item \code{id}: Identifier for each observation.
\item \code{trait}: Trait being evaluated.
\item \code{rank}: The rank assigned to each variety.
\item \code{variety}: The variety name.
}
}
\description{
This function transforms Tricot (Triadic Comparison of Technologies) data
into a tidy format, assigning ranks based on trait evaluations.
}
\details{
It supports input either as a single dataset (\code{data}) or as two datasets:
one with trait-wise rankings (\code{data_id_variety_trait}) and one with option labels and traits.

The function assumes a tricot setup where three options (e.g. crops, varieties) are ranked per trait.
Rankings may be extracted from columns ending in a suffix such as \verb{_pos} or \verb{_neg}, or explicitly provided.
}
\examples{
library(gosset)
# Read in tricot data from the `gosset` library
cassava_pivot <- pivot_tricot(cassava)

# Example using nicabean data - if we have ID-Variety-Trait Level data
data(nicabean)
nicabean_by_id_item_trait <- nicabean$trial
nicabean_by_id <- nicabean$covar 
nicabean_by_id_variety <- pivot_tricot(data_id_variety_trait = nicabean_by_id_item_trait, # B
                                       data_id_variety_trait_id_col = "id",
                                       variety_col = "item",
                                       trait_col = "trait",
                                       rank_col = "rank")

# Example using breadwheat data - if we have just ID-Level data 
data(breadwheat)
breadwheat_by_id_variety <- pivot_tricot(data = breadwheat,
                                         data_id_col = "participant_name",
                                         option_cols = c("variety_a", "variety_b", "variety_c"),
                                         trait_good = "best", trait_bad = "_worst")

# Example using nicabean data - if we have both data and data_id_variety_trait data
nicabean_by_id$Vigor_pos = "A"
nicabean_by_id$Vigor_neg = "C"
nicabean_by_id_variety_2 <- pivot_tricot(data = nicabean_by_id,
                                         data_id_col = "id",
                                         option_cols = c("variety_a", "variety_b", "variety_c"),
                                         data_id_variety_trait = nicabean_by_id_item_trait,
                                         data_id_variety_trait_id_col = "id",
                                         variety_col = "item",
                                         trait_col = "trait",
                                         rank_col = "rank")
}
